updated JourneyScreen with mapState, mapDispatch, and ComponentDidMount; created journey reducer; combined journey reducer in store
